+++ network_io/src/ZmqSocketSender.cpp
@@ -9,12 +9,12 @@ ZmqSocketSender::ZmqSocketSender(const std::string &endpoint) { m_endpoint = end
 void ZmqSocketSender::bind() {
     m_context = zmq_ctx_new();
     m_socket = zmq_socket(m_context, ZMQ_PUB);
-    int rc = zmq_bind(m_socket, m_endpoint.c_str());
+    size_t rc = zmq_bind(m_socket, m_endpoint.c_str());
     assert(rc == 0);
 }
 
-int ZmqSocketSender::send(ZmqHeader &header, const std::byte *data, size_t size, bool serialize_header) {
-    int rc;
+size_t ZmqSocketSender::send(ZmqHeader &header, const std::byte *data, size_t size, bool serialize_header) {
+    size_t rc;
     if (serialize_header) {
         rc = zmq_send(m_socket, &header, sizeof(ZmqHeader), ZMQ_SNDMORE);
         assert(rc == sizeof(ZmqHeader));
@@ -24,7 +24,7 @@ int ZmqSocketSender::send(ZmqHeader &header, const std::byte *data, size_t size,
         assert(rc == header_str.size());
     }
 
-    int rc2 = zmq_send(m_socket, data, size, 0);
+    size_t rc2 = zmq_send(m_socket, data, size, 0);
     assert(rc2 == size);
     return rc + rc2;
 }
